backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.32
pydantic==2.8.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
pymysql==1.1.1
python-dotenv==1.0.1
requests==2.32.3
email-validator==2.2.0
requests==2.32.3
mysql-connector-python==9.0.0
Pillow==10.4.0
pytesseract==0.3.10
python-multipart==0.0.9